pyproject.toml
[project]
name = "minedu-scrapper-identicole-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.115.5",
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.22",
]
